% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_stats.R
\name{cellProx.subgroup}
\alias{cellProx.subgroup}
\title{Analyze Proximity Subgroups}
\usage{
cellProx.subgroup(dt, number_of_simulations = 200)
}
\arguments{
\item{dt}{A data.table containing combined simulation results from multiple samples (must have `id`, `unified_int`, `round`, `V1` columns).}

\item{number_of_simulations}{Integer.}
}
\value{
A list from \code{cellProx.calcP}.
}
\description{
Aggregates results from multiple samples (rbind-ed tableResults) and calculates overall enrichment.
}
