% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_stats.R
\name{cellTriplet.sim}
\alias{cellTriplet.sim}
\title{Calculate Spatial Triplet Enrichment Using Permutations}
\usage{
cellTriplet.sim(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name,
  cluster_column,
  number_of_simulations = 200,
  min_cell_abundance = 0,
  specific_cell_types = NULL,
  pairwise_results_dt = NULL,
  pairwise_enrich_threshold = 1
)
}
\arguments{
\item{gobject}{A Giotto object.}

\item{spat_unit}{Spatial unit to use (default: NULL).}

\item{feat_type}{Feature type to use (default: NULL).}

\item{spatial_network_name}{Name of the spatial network to use.}

\item{cluster_column}{Column in cell metadata containing cluster annotations.}

\item{number_of_simulations}{Number of random simulations to perform (default: 200).}

\item{min_cell_abundance}{Minimum cells required to keep annotation group (default: 0).}

\item{specific_cell_types}{Character vector to isolate particular lineages.}

\item{pairwise_results_dt}{Core background summary data frame containing pairwise calculations.}

\item{pairwise_enrich_threshold}{Metric cutoff defining verified neighborhood structures.}
}
\value{
A data.table monitoring triplet formations across simulation trials.
}
\description{
Evaluates higher-order structural motifs (3-cliques / triangles) to characterize complex microenvironment niches.
}
