% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_stats.R
\name{defineNeighborhoods}
\alias{defineNeighborhoods}
\title{Define Neighborhood Proportions}
\usage{
defineNeighborhoods(annot.list)
}
\arguments{
\item{annot.list}{A list of annotated spatial networks.}
}
\value{
A list containing proportion tables, raw counts, and tracking sample IDs.
}
\description{
Calculates the proportion of cell types in the neighborhood of every cell.
}
