% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_stats.R
\name{exampleNeighborhoods.filter}
\alias{exampleNeighborhoods.filter}
\title{Filter Neighborhoods}
\usage{
exampleNeighborhoods.filter(res.list, indexCluster, chooseNeighbor, threshold)
}
\arguments{
\item{res.list}{Output from \code{defineNeighborhoods}.}

\item{indexCluster}{The cell type of the "center" cell to filter for.}

\item{chooseNeighbor}{The cell type of the neighbor to check proportions against.}

\item{threshold}{The minimum proportion required.}
}
\value{
A combined data.table of cells meeting the criteria.
}
\description{
Filters cells based on the proportion of a specific neighbor type.
}
