% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_halo.R
\name{makeGiottoObject}
\alias{makeGiottoObject}
\title{Create and Normalize Giotto Object from HALO Data}
\usage{
makeGiottoObject(inputFile, label, instructions = NULL)
}
\arguments{
\item{inputFile}{A raw data.table or data.frame from HALO.}

\item{label}{A character string for the sample name.}

\item{instructions}{A list of Giotto instructions (created via \code{Giotto::createGiottoInstructions}).}
}
\value{
A processed Giotto object.
}
\description{
Wraps the cleaning, creation, filtering, and normalization steps into one pipeline.
}
