% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{nicheContrast}
\alias{nicheContrast}
\title{Contrast Niche Composition Between Conditions}
\usage{
nicheContrast(
  annot.list,
  indexCluster,
  group.exp,
  group.ctrl,
  condition.exp = "Exp",
  condition.ctrl = "Ctrl",
  color_vector = NULL
)
}
\arguments{
\item{annot.list}{A list of annotated spatial network tables.}

\item{indexCluster}{Character string. The center cell type.}

\item{group.exp}{Character vector of sample IDs for the experimental group.}

\item{group.ctrl}{Character vector of sample IDs for the control group.}

\item{condition.exp}{Label for experimental group.}

\item{condition.ctrl}{Label for control group.}

\item{color_vector}{Named character vector of colors.}
}
\value{
A list containing the ggplot object (\code{scatterplot}) and the data table (\code{res}).
}
\description{
visualizes the difference in niche composition for a specific cell type between two experimental groups.
}
