% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{plot.globalChanges.Effectsize}
\alias{plot.globalChanges.Effectsize}
\title{Plot Global Effect Size Metrics as Violin Layouts}
\usage{
\method{plot}{globalChanges.Effectsize}(importIntxn.obj, colNames, title = "Stage 3 vs Stage 1")
}
\arguments{
\item{importIntxn.obj}{Comprehensive output object compiled via \code{findImportantInteractions}.}

\item{colNames}{String vector designating cell subset categories to track.}

\item{title}{Custom text heading to display over plot window layout.}
}
\value{
A ggplot visualization plot framework.
}
\description{
Evaluates and renders pooled delta shift distribution profiles across cell lineages.
}
