% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{plot.interaction.rankOrder}
\alias{plot.interaction.rankOrder}
\title{Generate Comprehensive Rank Order Effect Size Scatter Plots}
\usage{
\method{plot}{interaction.rankOrder}(
  imp.intxn.object,
  ef.threshold = 2,
  leaveOutClusters = NULL,
  labelSubset = NULL
)
}
\arguments{
\item{imp.intxn.object}{Tracking framework returned via \code{findImportantInteractions}.}

\item{ef.threshold}{Cutoff constraint isolating relevant target metrics rows.}

\item{leaveOutClusters}{String keys identifying noise structures to filter out from rendering view.}

\item{labelSubset}{String indicators isolating specific subsets to attach text annotations onto.}
}
\value{
A structured rank distribution plot framework.
}
\description{
Constructs clear distribution layouts tracking microenvironment perturbation shifts.
}
