% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{plotAverageNicheIgraph}
\alias{plotAverageNicheIgraph}
\title{Plot Average Niche as a Network Graph}
\usage{
plotAverageNicheIgraph(
  annot.list,
  indexCluster,
  title = NULL,
  color_vector = NULL,
  top_n = 8
)
}
\arguments{
\item{annot.list}{A list of annotated spatial network tables.}

\item{indexCluster}{Character string. The center cell type.}

\item{title}{Character string.}

\item{color_vector}{Named character vector of colors.}

\item{top_n}{Integer. Number of top neighbors to display (default: 8).}
}
\value{
A ggraph object.
}
\description{
Visualizes the top neighbors of a specific cell type as a star-like network graph using `ggraph`.
}
