% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{plotNeighborHistogram}
\alias{plotNeighborHistogram}
\title{Plot Neighbor-Dependent Feature Expression}
\usage{
plotNeighborHistogram(
  gobject,
  neighborhood_prop_table,
  target_cluster,
  neighbor_cluster,
  feature,
  title = NULL
)
}
\arguments{
\item{gobject}{A Giotto object.}

\item{neighborhood_prop_table}{Data.table from `defineNeighborhoods` (one sample).}

\item{target_cluster}{The cell type of the "center" cell.}

\item{neighbor_cluster}{The neighbor cell type to stratify by.}

\item{feature}{Gene or protein name to plot.}

\item{title}{Plot title.}
}
\value{
A list containing the plot and the p-value.
}
\description{
Compares feature expression in cells with neighbors of type X vs those without.
Requires a Giotto object to retrieve expression data.
}
