% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_viz.R
\name{plotInteractionEffectSizes}
\alias{plotInteractionEffectSizes}
\title{Plot Interaction Effect Sizes}
\usage{
plotInteractionEffectSizes(df, condition.exp, condition.ctrl)
}
\arguments{
\item{df}{Data.table output from \code{calculateInteractionEffectSizes}.}

\item{condition.exp}{Label for experimental condition.}

\item{condition.ctrl}{Label for control condition.}
}
\value{
A ggplot object.
}
\description{
Visualizes the output of \code{calculateInteractionEffectSizes} as a bar chart.
}
